+++ src/main/java/com/richashworth/planningpoker/model/Estimate.java
@@ -10,5 +10,5 @@ import lombok.Data;
 @AllArgsConstructor
 public class Estimate {
     private String userId;
-    private int estimateValue;
+    private double estimateValue;
 }
